# week-05/report_assets/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "3D Printer Production Simulator",
        "description": "3D Printer Production Simulator API — manage inventory, orders, and production cycles.",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "List Orders",
                "description": "List all manufacturing orders (newest first).",
                "operationId": "list_orders_api_orders_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/OrderResponse"
                                    },
                                    "type": "array",
                                    "title": "Response List Orders Api Orders Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Create Order",
                "description": "Create a new manufacturing order manually.",
                "operationId": "create_order_api_orders_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
        "/api/orders/{order_id}": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Get Order",
                "description": "Get order details including BOM breakdown.",
                "operationId": "get_order_api_orders__order_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Order Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/orders/{order_id}/cancel": {
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Cancel Order",
                "description": "Cancel an order and release any reserved materials.",
                "operationId": "cancel_order_api_orders__order_id__cancel_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Order Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/purchase-orders": {
            "get": {
                "tags": [
                    "purchase-orders"
                ],
                "summary": "List Purchase Orders",
                "description": "List all purchase orders.",
                "operationId": "list_purchase_orders_api_purchase_orders_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/POResponse"
                                    },
                                    "type": "array",
                                    "title": "Response List Purchase Orders Api Purchase Orders Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "purchase-orders"
                ],
                "summary": "Create Purchase Order",
                "description": "Create a purchase order. Applies bulk discount tiers automatically.",
                "operationId": "create_purchase_order_api_purchase_orders_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePORequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/POResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/purchase-orders/{po_id}": {
            "get": {
                "tags": [
                    "purchase-orders"
                ],
                "summary": "Get Purchase Order",
                "description": "Get purchase order details.",
                "operationId": "get_purchase_order_api_purchase_orders__po_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "po_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Po Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/POResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/purchase-orders/{po_id}/cancel": {
            "post": {
                "tags": [
                    "purchase-orders"
                ],
                "summary": "Cancel Purchase Order",
                "description": "Cancel a pending purchase order.",
                "operationId": "cancel_purchase_order_api_purchase_orders__po_id__cancel_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "po_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Po Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateOrderRequest": {
                "properties": {
                    "product_model": {
                        "type": "string",
                        "title": "Product Model"
                    },
                    "quantity": {
                        "type": "number",
                        "title": "Quantity"
                    }
                },
                "type": "object",
                "required": [
                    "product_model",
                    "quantity"
                ],
                "title": "CreateOrderRequest"
            },
            "CreatePORequest": {
                "properties": {
                    "supplier_id": {
                        "type": "integer",
                        "title": "Supplier Id"
                    },
                    "product_name": {
                        "type": "string",
                        "title": "Product Name"
                    },
                    "quantity": {
                        "type": "number",
                        "title": "Quantity"
                    }
                },
                "type": "object",
                "required": [
                    "supplier_id",
                    "product_name",
                    "quantity"
                ],
                "title": "CreatePORequest"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "OrderResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "product_model": {
                        "type": "string",
                        "title": "Product Model"
                    },
                    "quantity_needed": {
                        "type": "number",
                        "title": "Quantity Needed"
                    },
                    "quantity_produced": {
                        "type": "number",
                        "title": "Quantity Produced"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "created_date": {
                        "type": "string",
                        "title": "Created Date"
                    },
                    "started_date": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Started Date"
                    },
                    "completed_date": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Completed Date"
                    },
                    "failure_reason": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Failure Reason"
                    },
                    "bom_requirements": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bom Requirements"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "product_model",
                    "quantity_needed",
                    "quantity_produced",
                    "status",
                    "created_date"
                ],
                "title": "OrderResponse"
            },
            "POResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "supplier_id": {
                        "type": "integer",
                        "title": "Supplier Id"
                    },
                    "supplier_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Supplier Name"
                    },
                    "product_name": {
                        "type": "string",
                        "title": "Product Name"
                    },
                    "quantity_ordered": {
                        "type": "number",
                        "title": "Quantity Ordered"
                    },
                    "quantity_delivered": {
                        "type": "number",
                        "title": "Quantity Delivered"
                    },
                    "unit_cost": {
                        "type": "number",
                        "title": "Unit Cost"
                    },
                    "total_cost": {
                        "type": "number",
                        "title": "Total Cost"
                    },
                    "order_date": {
                        "type": "string",
                        "title": "Order Date"
                    },
                    "expected_delivery": {
                        "type": "string",
                        "title": "Expected Delivery"
                    },
                    "actual_delivery": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Actual Delivery"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "supplier_id",
                    "product_name",
                    "quantity_ordered",
                    "quantity_delivered",
                    "unit_cost",
                    "total_cost",
                    "order_date",
                    "expected_delivery",
                    "status"
                ],
                "title": "POResponse"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "api/auth/login"
                    }
                }
            }
        }
    }
}